Rename css_parser module -> style_properties
